[s2750] Implement script for Chess Event - Karazhan
Please note that this is only a basic implementation. The AI works fine, but it's not as good as the AI on retail. There are some fine details which can be improved. Also some spells may require core fixes, please check the forum for details.

If you have any suggestions or bugs to report, please feel free to contact us on github, scriptdev2 or cmangos forums.

Finally I want to give a very special thanks to X-Savior, evil@wow, YuruY, cala and Schmoozerd for the data and support provided.

(based on commit f995b94)

Signed-off-by: Xfurry <xfurry@scriptdev2.com>