[s2852] Fix quest 1249, the missing diplo_pt11
Sidenote: Due to changes in escortAI, checking for motion-tpye in MovementInform must be checked in every case.
Added wrapper for this to be prepared for easier transition
Also note that when Reaching the COMBAT_START point for escortAI will no more trigger MovementInform.

Thanks to xfurry for researching, as well as zimms in unified-db/Database~74 for reporting and all others who contributed to the problem research!

(based on commit SD2-TBC[3150] - fcdfa6d)

Signed-off-by: xfurry <xfurry@scriptdev2.com>